+++ CHANGELOG.md
@@ -15,10 +15,11 @@ and this project adheres to [Semantic Versioning][].
 ### Changed
 
 -   `obst` and `vart` create local copy of `nx.DiGraphs` that are added (#26)
+-   `TreeData.label` value remains the same after `td.concat` as long as all `label` values are the same (#27)
 
 ### Fixed
 
--   Fixed bug which caused key to be listed twice in `tree_label` column after value update in `obst` or `vart` (#26)
+-   Fixed bug which caused key to be listed twice in `label` column after value update in `obst` or `vart` (#26)
 
 ## [0.0.2] - 2024-06-18